+++ common/src/main/java/io/github/mortuusars/envelope/command/EnvelopeCommand.java
@@ -1,0 +1,156 @@
+package io.github.mortuusars.envelope.command;
+
+import com.mojang.brigadier.CommandDispatcher;
+import com.mojang.brigadier.context.CommandContext;
+import com.mojang.brigadier.exceptions.CommandSyntaxException;
+import io.github.mortuusars.envelope.command.argument.AddressArgument;
+import io.github.mortuusars.envelope.command.suggestion.AddressSuggestions;
+import io.github.mortuusars.envelope.world.delivery.BackgroundCourier;
+import io.github.mortuusars.envelope.world.mail.address.Address;
+import io.github.mortuusars.envelope.world.mail.address.AddressDisplay;
+import io.github.mortuusars.envelope.world.mail.address.validation.AddressValidator;
+import io.github.mortuusars.envelope.world.mail.address.validation.Validator;
+import net.minecraft.ChatFormatting;
+import net.minecraft.commands.CommandBuildContext;
+import net.minecraft.commands.CommandSourceStack;
+import net.minecraft.commands.Commands;
+import net.minecraft.commands.arguments.item.ItemArgument;
+import net.minecraft.commands.arguments.item.ItemInput;
+import net.minecraft.core.BlockPos;
+import net.minecraft.network.chat.*;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.item.ItemStack;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
+
+public class EnvelopeCommand {
+    public static void register(CommandDispatcher<CommandSourceStack> dispatcher, CommandBuildContext context) {
+        dispatcher.register(Commands.literal("envelope")
+              .requires((stack) -> stack.hasPermission(2))
+              .then(Commands.literal("send")
+                    .then(Commands.argument("mail", ItemArgument.item(context))
+                          .executes(c -> sendMail(c, ItemArgument.getItem(c, "mail")))))
+              .then(Commands.literal("pigeonhole")
+                    .then(Commands.literal("list")
+                          .executes(EnvelopeCommand::listAllPigeonholes))
+                    .then(Commands.literal("position")
+                          .then(Commands.argument("address", AddressArgument.pigeonhole())
+                                .suggests(AddressSuggestions.pigeonhole())
+                                .executes(c -> pigeonholePosition(c, AddressArgument.getPigeonhole(c, "address")))))
+                    .then(Commands.literal("rename")
+                          .then(Commands.argument("address", AddressArgument.pigeonhole())
+                                .suggests(AddressSuggestions.pigeonhole())
+                                .then(Commands.argument("new_address", AddressArgument.all())
+                                      .executes(c -> renamePigeonhole(c,
+                                            AddressArgument.getPigeonhole(c, "address"),
+                                            AddressArgument.getPigeonhole(c, "new_address"))))))));
+    }
+
+    // -- Mail
+
+    private static int sendMail(CommandContext<CommandSourceStack> context, ItemInput item) throws CommandSyntaxException {
+        ServerLevel level = context.getSource().getLevel();
+        ItemStack mail = item.createItemStack(1, false);
+
+        try {
+            BackgroundCourier courier = BackgroundCourier.virtual();
+            courier.startDelivery(level, mail);
+            level.getEnvelopeContext().getBackgroundDelivery().add(courier);
+            Component senderName = courier.getDelivery()
+                  .map(d -> d.getSenderAddress().getDisplayName().withStyle(Style.EMPTY.withColor(AddressDisplay.SENDER_COLOR)))
+                  .orElse(Component.empty());
+            context.getSource().sendSuccess(() ->
+                  Component.literal("Mail sent to ").append(senderName), true);
+        } catch (Exception e) {
+            context.getSource().sendFailure(Component.literal("Cannot send: " + e.getMessage()));
+        }
+
+        return 0;
+    }
+
+    // -- Pigeonhole
+
+    private static int listAllPigeonholes(CommandContext<CommandSourceStack> context) {
+        ServerLevel level = context.getSource().getLevel();
+        Set<Address.Pigeonhole> addresses = level.getEnvelopeContext().getPigeonholeManager().getAllAddresses();
+
+        if (!addresses.isEmpty()) {
+            context.getSource().sendSuccess(() -> Component.literal("All pigeonholes in this dimension:"), true);
+            for (Address.Pigeonhole address : addresses) {
+                context.getSource().sendSuccess(() -> copyableAddressAndPos(address,
+                      level.getEnvelopeContext().getPigeonholeManager().getPositionOf(address)), true);
+            }
+        } else {
+            context.getSource().sendSuccess(() ->
+                  Component.literal("There are no pigeonholes in this dimension."), true);
+        }
+        return 0;
+    }
+
+    private static int pigeonholePosition(CommandContext<CommandSourceStack> context, Address.Pigeonhole address) {
+        ServerLevel level = context.getSource().getLevel();
+        if (!level.getEnvelopeContext().getPigeonholeManager().exists(address)) {
+            context.getSource().sendFailure(address.getDisplayName().append(" does not exist."));
+            return 1;
+        }
+
+        level.getEnvelopeContext().getPigeonholeManager().getPositionOf(address)
+              .ifPresentOrElse(
+                    pos -> context.getSource().sendSuccess(() -> copyableAddressAndPos(address, Optional.of(pos)), true),
+                    () -> context.getSource().sendFailure(address.getDisplayName()
+                          .append(" does not have a position associated with it.")));
+        return 0;
+    }
+
+    private static int renamePigeonhole(CommandContext<CommandSourceStack> context,
+                                        Address.Pigeonhole oldAddress, Address.Pigeonhole newAddress) {
+        ServerLevel level = context.getSource().getLevel();
+
+        if (!level.getEnvelopeContext().getPigeonholeManager().exists(oldAddress)) {
+            context.getSource().sendFailure(Component.literal("Address '"
+                  + oldAddress.getDisplayName().getString() + "' does not exist."));
+            return 1;
+        }
+
+        if (oldAddress.equals(newAddress)) {
+            context.getSource().sendFailure(Component.literal("Addresses are the same.").withStyle(ChatFormatting.RED));
+            return 1;
+        }
+
+        List<Validator.Issue> validate = new AddressValidator().setKnownAddresses(level.getEnvelopeContext().getKnownAddresses())
+              .validate(newAddress.id());
+        if (!validate.isEmpty()) {
+            context.getSource().sendFailure(Component.literal("Invalid address: ").append(validate.getFirst().translate()));
+            return 1;
+        }
+
+        level.getEnvelopeContext().getPigeonholeManager().rename(oldAddress, newAddress);
+
+        context.getSource().sendSuccess(() -> Component.literal(
+              oldAddress.getDisplayName().getString() + " has been renamed to " + newAddress.id() + "."), true);
+
+        return 0;
+    }
+
+    private static MutableComponent copyableAddressAndPos(Address address, Optional<BlockPos> pos) {
+        String addressId = address.getDisplayName().getString();
+        String posStr = pos.map(BlockPos::toShortString).orElse("");
+        String posToCopy = posStr.replace(",", "");
+
+        return address.getDisplayName()
+              .withStyle(Style.EMPTY
+                    .withColor(AddressDisplay.GENERIC_COLOR)
+                    .withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, Component.literal("Copy Address")
+                          .append("\n")
+                          .append(Component.literal(addressId).withStyle(ChatFormatting.GRAY))))
+                    .withClickEvent(new ClickEvent(ClickEvent.Action.COPY_TO_CLIPBOARD, addressId)))
+              .append(Component.literal("@[" + posStr + "]").withStyle(Style.EMPTY
+                    .withColor(ChatFormatting.WHITE)
+                    .withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, Component.literal("Copy Position")
+                          .append("\n")
+                          .append(Component.literal(posToCopy).withStyle(ChatFormatting.GRAY))))
+                    .withClickEvent(new ClickEvent(ClickEvent.Action.COPY_TO_CLIPBOARD, posToCopy))));
+    }
+}