+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 0.2.0.2
+- Changed command `/mail send [item]` to `/envelope send [item]`
+- Added `/envelope pigeonhole [list|position|rename]`  commands
 - Improvements to Bugger (Debug utility enabled with `debug.debug` server config option. Adds more info to F3 screen, more log messages, etc).
 
 ## 0.2.0.1